Updated build files for pca10001, pca10028 and pca10040
The build files for pca10028 and pca10040 closely resemble the ones used by the
ble_app_beacon example from nRF5_SDK_12.3.0_d7731ad, which should make them
easier to update after future changes to the SDK.

The build files for pca10001 are identical to pca10028 except for the board name
and the RAM area.

pca10040 does not have the ADC peripheral, and battery levels must be fetched
using SAADC instead. This is not implemented, and we currently hardcode the
battery level to 100 when no ADC is present.